+++ R/zzz.R
@@ -4,6 +4,6 @@
   packageStartupMessage("For more information, type ?lavinteract.")
   packageStartupMessage(
     "For suggestions or to report issues, please contact Giuseppe Corbelli at ",
-    "giuseppe.corbelli@uniroma1.it or giuseppe.corbelli@uninettunouniversity.net."
+    "giuseppe.corbelli@uninettunouniversity.net."
   )
 }